requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
djangorestframework==3.12.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
djangorestframework-simplejwt
autopep8